Reorder files
